requirements.txt
streamlit
openai>=1.0.0
python-docx
PyPDF2
spacy
fuzzywuzzy
python-dotenv
